refactor(postgres): centralize management authority
