+++ crates/automata-ci-postgres/src/auth/management/runner_enrollment.rs
@@ -315,6 +315,7 @@ impl PostgresHumanRbacManagementRepository {
             &request.actor,
             &["runners:enroll"],
             descriptor,
+            map_database_error,
         )
         .await?;
         let MutationAuthorization::Authorized(actor) = authorization else {
